# ENH: Faster zero-phase filtering (and better tests)