This code rewrites the overlap-add filtering to use one pass, whether the filtering is zero-phase or not. It does that by allowing the impulse response to be acausal (which is done by taking the magnitude-squared of `h_fft`) and handling the overlap on both sides of each segment, instead of just the right side.

Here is a simple speed test on the current master:

```
<RawFIF  |  n_channels x n_times : 401 x 379200>
filter length: 1024
    n_jobs=1,  time: 12.1 s
    n_jobs=cuda,  time: 12.0 s
filter length: 4096
    n_jobs=1,  time: 13.5 s
    n_jobs=cuda,  time: 5.6 s
filter length: 16384
    n_jobs=1,  time: 18.2 s
    n_jobs=cuda,  time: 6.7 s
filter length: 65536
    n_jobs=1,  time: 30.3 s
    n_jobs=cuda,  time: 7.8 s
```

And on the proposed branch:
```
<RawFIF  |  n_channels x n_times : 401 x 379200>
filter length: 1024
    n_jobs=1,  time: 6.4 s
    n_jobs=cuda,  time: 4.0 s
filter length: 4096
    n_jobs=1,  time: 7.2 s
    n_jobs=cuda,  time: 2.9 s
filter length: 16384
    n_jobs=1,  time: 12.2 s
    n_jobs=cuda,  time: 3.8 s
filter length: 65536
    n_jobs=1,  time: 22.0 s
    n_jobs=cuda,  time: 5.6 s
```

Speedup is around 2x under some conditions, which matches the theoretical expectations. This change also warrants re-examining the choice of segment length for the overlap-add, but that is for a different PR.

In addition to helping with the filtering code, @Eric89GXL also added much better tests of the filtering code.

I think I did some dumb stuff like merging master into the PR a couple times (sorry) but from looking at the changed files it looks like maybe that doesn't matter? I can withdraw PR and fix things if that makes it cleaner.