@Eric89GXL 2015-07-06
Rebase / cleaning appears to have worked, and tests pass over here. LGTM +1 for merge, but I helped develop the code so I am biased :)