@agramfort 2015-07-07
please update what's new and then merge. LGTM

thanks @rkmaddox and @Eric89GXL !