@Eric89GXL 2015-07-07
Ross in case you're curious 

https://github.com/mne-tools/mne-python/commit/9714cf60cc14aa172b3591f37409176aa6730a53